Wrap client login in suspense
